Proper representation of feature/weight vector and CRF.
The CRF is now a list of 0s with len == the sentence. Max-sum BP will
update the CRF's values as we pass messages and decode tags.

More importantly, the feature/weight vector now has a correct
representation. Each feature has a # of tags x |feature| ndarray
representing the # of times each tag had each feature value. This fixes
the previous implementation, where the feature vector was a humongous
flat ndarray.

With the above patch to the feature vector, we also now generate node
and edge potentials. This means we can start passing messages in the
next commit.